+++ core/Cargo.toml
@@ -4,14 +4,11 @@ version = "0.10.0"
 edition = "2018"
 
 [dependencies]
-chrono = "0.4"
 futures = "0.1.21"
 graph = { path = "../graph" }
 graph-graphql = { path = "../graphql" }
 graph-runtime-wasm = { path = "../runtime/wasm" }
-itertools = "0.7"
 lazy_static = "1.2.0"
-reqwest = "0.9"
 semver = "0.9.0"
 serde = "1.0"
 serde_json = "1.0"
