+++ core/src/lib.rs
@@ -1,12 +1,9 @@
-extern crate chrono;
 extern crate futures;
 extern crate graph;
 extern crate graph_graphql;
 #[cfg(test)]
 extern crate graph_mock;
 extern crate graph_runtime_wasm;
-extern crate itertools;
-extern crate reqwest;
 extern crate serde;
 #[macro_use]
 extern crate serde_json;
@@ -16,12 +13,9 @@ extern crate serde_yaml;
 extern crate uuid;
 
 mod graphql;
-mod log;
 mod subgraph;
 
 pub use crate::graphql::GraphQlRunner;
-pub use crate::log::elastic::{elastic_logger, ElasticDrainConfig, ElasticLoggingConfig};
-pub use crate::log::split::split_logger;
 pub use crate::subgraph::{
     DataSourceLoader, SubgraphAssignmentProvider, SubgraphInstanceManager, SubgraphRegistrar,
 };
