--- core/src/log/mod.rs
@@ -1,2 +1,0 @@
-pub mod elastic;
-pub mod split;
