+++ graph/Cargo.toml
@@ -8,7 +8,9 @@ backtrace = "0.3.9"
 bigdecimal = { version = "0.0.14", features = ["serde"] }
 bytes = "0.4.11"
 diesel = { version = "1.3.2", features = ["postgres", "serde_json", "numeric", "r2d2"] }
-
+chrono = "0.4"
+itertools = "0.7"
+reqwest = "0.9"
 # We would like to switch to upstream master once https://github.com/paritytech/ethabi/pull/145
 # is merged.
 # But, graph-patches also contains Jannis' https://github.com/paritytech/ethabi/pull/140 PR
