+++ graph/src/util/mod.rs
@@ -1,9 +1,6 @@
 /// Utilities for working with futures.
 pub mod futures;
 
-/// Logging facilities based on `slog`.
-pub mod log;
-
 /// Utils for working with ethereum data types
 pub mod ethereum;
